+++ controllers/WebsiteController.php
@@ -21,33 +21,38 @@ use humhub\modules\user\models\Group;
 
 
 /**
- * When Humhub is host
+ * When Humhub is host (if guest, redirects to external website)
  * Show external website pages in an iframe and pages contents addons beside
+ * @param $id Website ID
+ * @param $pageId Page ID
+ * @param $pageUrl string Page URL
  */
 class WebsiteController extends ContentContainerController
 {
-    public function actionIndex ($id)
+    public function actionIndex ($id, $pageId = null, $pageUrl = null)
     {
-        // Get website from title
+        // Get website
         $website = Website::findOne($id);
         if ($website === null) {
             throw new HttpException(404);
         }
 
-        // Set first page URL
-        $pageUrl = $website->first_page_url;
-
-        // If pageId is in the URL
-        $pageId = Yii::$app->request->get('pageId');
+        // If $pageId not null and Page exists, set pageUrl from page
         if ($pageId !== null) {
             $page = Page::findOne($pageId);
             if ($page !== null) {
                 $pageUrl = $page->page_url;
             }
         }
-        // If pageUrl is in the URL (Page does not exist)
-        else {
-            $pageUrl = Yii::$app->request->get('pageUrl', $pageUrl);
+        // If pageUrl is null
+        elseif (empty($pageUrl)) {
+            // Set first page URL
+            $pageUrl = $website->first_page_url;
+        }
+
+        // If Humhub is guest, redirect to external website
+        if (!$website->humhub_is_host) {
+            return $this->redirect($pageUrl);
         }
 
         return $this->render('index', [