+++ messages/ko/model.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  'Filter' => '필터',
+);