+++ messages/ko/models.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  '{title}: show comments' => '{title} : 댓글 표시',
+);