+++ messages/vi/models.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  '{title}: show comments' => '{title} : hiển thị bình luận',
+);