+++ docs/CHANGELOG.md
@@ -4,10 +4,12 @@ Changelog
 V0.20 (November, 25, 2020)
 --------------------
 - Chg: Module renamed
-- Chg: script in the external website if guest (integrated in iframe): `url` is now `pageUrl` and `title` is now `pageTitle`
+- Chg: Script in the external website if guest (integrated in iframe): `url` is now `pageUrl` and `title` is now `pageTitle`
 - Chg: Changed name of the tables, files, models, controller and view
 - Chg: Changed `FirstCommentForm` widget to new Hummhub 1.7 specifications
 - Enh: Added auto login with SSO
+- Enh: Wall entry updated for Humhub 1.7
+- Enh: If the content related to a page is archived and all comments have been removed, only the permalink will be shown
 
 
 V0.15 (November, 23, 2020)